rename: questio.type 파일명 수정에 따른 변경
